moved and cleaned up the file and class structure. compiling and working demo
