+++ brisk/CMakeLists.txt
@@ -17,12 +17,21 @@ set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
 set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
 
 add_definitions(-msse2 -mssse3 -march=native -Wextra -Wall
-                -Wno-overloaded-virtual -Wno-non-virtual-dtor #OpenCV
-                )
+                -Wno-overloaded-virtual -Wno-non-virtual-dtor # OpenCV.
+                -std=c++11)
 
-rosbuild_add_library(${PROJECT_NAME} src/brisk.cpp
-                                     src/harris.cpp 
-                                     src/HarrisScoreCalculator.cpp 
-                                     src/HarrisScoreCalculatorFloat.cpp 
-                                     src/harrisScores_sse.cpp 
-                                     src/rdtsc_wrapper.cpp)
+rosbuild_add_library(${PROJECT_NAME} src/brisk-descriptor-extractor.cc  
+                                     src/brisk-feature-detector.cc   
+                                     src/brisk-layer.cc
+                                     src/brisk-scale-space.cc
+                                     src/brute-force-matcher.cc   
+                                     src/harris-feature-detector.cc
+                                     src/harris-score-calculator.cc
+                                     src/harris-score-calculator-float.cc
+                                     src/harris-scores-sse.cc
+                                     src/rdtsc-wrapper.cc
+                                     src/sse-filters.cc
+)
+
+rosbuild_add_executable(demo src/demo.cc)
+target_link_libraries(demo ${PROJECT_NAME})
