+++ brisk/include/brisk/brute-force-matcher.h
@@ -38,13 +38,42 @@
  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef BRISK_BRISK_H_
-#define BRISK_BRISK_H_
+#ifndef BRISK_BRUTE_FORCE_MATCHER_H_
+#define BRISK_BRUTE_FORCE_MATCHER_H_
 
-#include <brisk/brisk-descriptor-extractor.h>
-#include <brisk/brisk-feature-detector.h>
-#include <brisk/harris-feature-detector.h>
-#include <brisk/harris-score-calculator.h>
-#include <brisk/scale-space-feature-detector.h>
+#include <brisk/brisk-opencv.h>
+#include <brisk/internal/hamming-sse.h>
+#include <brisk/internal/macros.h>
 
-#endif  //BRISK_BRISK_H_
+namespace cv {
+class CV_EXPORTS BruteForceMatcherSse : public DescriptorMatcher {
+public:
+  BruteForceMatcherSse(
+      const brisk::HammingSse& distance = brisk::HammingSse()) :
+        distance_(distance) { }
+  virtual ~BruteForceMatcherSse() { }
+  virtual bool isMaskSupported() const { return true; }
+  virtual cv::Ptr<DescriptorMatcher> clone(bool emptyTrainData = false ) const;
+
+protected:
+  virtual void knnMatchImpl(const Mat& queryDescriptors,
+                            vector<vector<DMatch> >& matches, int k,
+      const vector<Mat>& masks=vector<Mat>(), bool compactResult = false );
+  virtual void radiusMatchImpl(const Mat& queryDescriptors,
+                               vector<vector<DMatch> >& matches,
+                               float maxDistance,
+      const vector<Mat>& masks = vector<Mat>(), bool compactResult = false );
+
+  brisk::HammingSse distance_;
+
+private:
+  //  Next two methods are used to implement specialization.
+  static void commonKnnMatchImpl( BruteForceMatcherSse& matcher,
+      const Mat& queryDescriptors, vector<vector<DMatch> >& matches, int k,
+      const vector<Mat>& masks, bool compactResult );
+  static void commonRadiusMatchImpl( BruteForceMatcherSse& matcher,
+      const Mat& queryDescriptors, vector<vector<DMatch> >& matches,
+      float maxDistance, const vector<Mat>& masks, bool compactResult );
+};
+}  // namespace cv
+#endif  // BRISK_BRUTE_FORCE_MATCHER_H_
