+++ brisk/include/brisk/internal/harris-scores-sse.h
@@ -38,13 +38,15 @@
  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef BRISK_BRISK_H_
-#define BRISK_BRISK_H_
+#ifndef BRISK_INTERNAL_HARRIS_SCORES_SSE_HPP_
+#define BRISK_INTERNAL_HARRIS_SCORES_SSE_HPP_
 
-#include <brisk/brisk-descriptor-extractor.h>
-#include <brisk/brisk-feature-detector.h>
-#include <brisk/harris-feature-detector.h>
-#include <brisk/harris-score-calculator.h>
-#include <brisk/scale-space-feature-detector.h>
+#include <opencv2/opencv.hpp>
+#include <brisk/brisk.h>
 
-#endif  //BRISK_BRISK_H_
+namespace brisk {
+// SSE speeded up (dxdx dxdy and dydy only). based on harrisScores_basic_noMats(.) .
+void harrisScores_sse(const cv::Mat& src, cv::Mat& scores);
+
+}  // namespace brisk
+#endif   // BRISK_INTERNAL_HARRIS_SCORES_SSE_HPP_
