+++ brisk/include/brisk/internal/helper-structures.h
@@ -38,13 +38,25 @@
  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef BRISK_BRISK_H_
-#define BRISK_BRISK_H_
+#ifndef BRISK_INTERNAL_HELPER_STRUCTURES_H_
+#define BRISK_INTERNAL_HELPER_STRUCTURES_H_
 
-#include <brisk/brisk-descriptor-extractor.h>
-#include <brisk/brisk-feature-detector.h>
-#include <brisk/harris-feature-detector.h>
-#include <brisk/harris-score-calculator.h>
-#include <brisk/scale-space-feature-detector.h>
-
-#endif  //BRISK_BRISK_H_
+namespace brisk {
+// Some helper structures for the Brisk pattern representation.
+struct BriskPatternPoint {
+  float x;         // x coordinate relative to center.
+  float y;         // x coordinate relative to center.
+  float sigma;     // Gaussian smoothing sigma.
+};
+struct BriskShortPair {
+  unsigned int i;  // Index of the first pattern point.
+  unsigned int j;  // Index of other pattern point.
+};
+struct BriskLongPair {
+  unsigned int i;  // Index of the first pattern point.
+  unsigned int j;  // Index of other pattern point.
+  int weighted_dx;  // 1024.0 / dx.
+  int weighted_dy;  // 1024.0 / dy.
+};
+}  // namespace brisk
+#endif  // BRISK_INTERNAL_HELPER_STRUCTURES_H_
