+++ brisk/include/brisk/internal/scale-space-layer-inl.h
@@ -1,13 +1,49 @@
 /*
- * ScaleSpaceLayer.hpp
- *
- *  Created on: Aug 5, 2012
- *      Author: lestefan
+ Copyright (C) 2011  The Autonomous Systems Lab, ETH Zurich,
+ Stefan Leutenegger, Simon Lynen and Margarita Chli.
+
+ Copyright (C) 2013  The Autonomous Systems Lab, ETH Zurich,
+ Stefan Leutenegger and Simon Lynen.
+
+ BRISK - Binary Robust Invariant Scalable Keypoints
+ Reference implementation of
+ [1] Stefan Leutenegger,Margarita Chli and Roland Siegwart, BRISK:
+ Binary Robust Invariant Scalable Keypoints, in Proceedings of
+ the IEEE International Conference on Computer Vision (ICCV2011).
+
+ This file is part of BRISK.
+
+ All rights reserved.
+
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions are met:
+     * Redistributions of source code must retain the above copyright
+       notice, this list of conditions and the following disclaimer.
+     * Redistributions in binary form must reproduce the above copyright
+       notice, this list of conditions and the following disclaimer in the
+       documentation and/or other materials provided with the distribution.
+     * Neither the name of the <organization> nor the
+       names of its contributors may be used to endorse or promote products
+       derived from this software without specific prior written permission.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+ DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <sys/time.h>
-#include <brisk/rdtsc_wrapper.h>
+#ifndef BRISK_INTERNAL_SCALE_SPACE_LAYER_INL_H_
+#define BRISK_INTERNAL_SCALE_SPACE_LAYER_INL_H_
 
+#include <brisk/internal/rdtsc-wrapper.h>
+
+namespace brisk {
 template<class SCORE_CALCULTAOR_T>
 ScaleSpaceLayer<SCORE_CALCULTAOR_T>::ScaleSpaceLayer(const cv::Mat& img,
                                                      bool initScores) {
@@ -67,7 +103,7 @@ template<class SCORE_CALCULTAOR_T>
 void ScaleSpaceLayer<SCORE_CALCULTAOR_T>::create(
     ScaleSpaceLayer<ScoreCalculator_t>* layerBelow, bool initScores) {
   // For successive construction.
-  rdtsc::timing::Timer timerFancyOp0(
+  rdtsc::timing::Timer timerDownsample(
       "0.0 BRISK Detection: Creation&Downsampling (per layer)");
   int type = layerBelow->_img.type();
   if (layerBelow->_isOctave) {
@@ -115,7 +151,7 @@ void ScaleSpaceLayer<SCORE_CALCULTAOR_T>::create(
     _scale = pow(2.0, double(_layerNumber / 2)) * 1.5;
     _offset = _scale * 0.5 - 0.5;
   }
-  timerFancyOp0.stop();
+  timerDownsample.stop();
 
   // By default no uniformity radius.
   _radius = 1;
@@ -168,10 +204,10 @@ void ScaleSpaceLayer<SCORE_CALCULATOR_T>::detectScaleSpaceMaxima(
       }
     }
   } else {
-    TimerSwitchable timerFancyOp1(
+    TimerSwitchable timerNonMaxSuppression2d(
         "0.2 BRISK Detection: 2d nonmax suppression (per layer)");
     _scoreCalculator.get2dMaxima(points, _absoluteThreshold);
-    timerFancyOp1.stop();
+    timerNonMaxSuppression2d.stop();
   }
   // Next check above and below. The code looks a bit stupid, but that's
   // for speed. We don't want to make the distinction analyzing whether or
@@ -179,7 +215,7 @@ void ScaleSpaceLayer<SCORE_CALCULATOR_T>::detectScaleSpaceMaxima(
   if (!usePassedKeypoints) {
     if (_aboveLayer_ptr != 0 && _belowLayer_ptr != 0) {
       // Check above and below
-      TimerSwitchable timerFancyOp2(
+      TimerSwitchable timerNonMaxSuppression3d(
           "0.3 BRISK Detection: 3d nonmax suppression (per layer)");
       std::vector<typename ScoreCalculator_t::PointWithScore> pt_tmp;
       pt_tmp.reserve(points.size());
@@ -239,10 +275,10 @@ void ScaleSpaceLayer<SCORE_CALCULATOR_T>::detectScaleSpaceMaxima(
         pt_tmp.push_back(*it);
       }
       points.assign(pt_tmp.begin(), pt_tmp.end());
-      timerFancyOp2.stop();
+      timerNonMaxSuppression3d.stop();
     } else if (_aboveLayer_ptr != 0) {
       // Check above.
-      TimerSwitchable timerFancyOp2(
+      TimerSwitchable timerNonMaxSuppression3d(
           "0.3 BRISK Detection: 3d nonmax suppression (per layer)");
       std::vector<typename ScoreCalculator_t::PointWithScore> pt_tmp;
       pt_tmp.reserve(points.size());
@@ -280,10 +316,10 @@ void ScaleSpaceLayer<SCORE_CALCULATOR_T>::detectScaleSpaceMaxima(
 
       }
       points.assign(pt_tmp.begin(), pt_tmp.end());
-      timerFancyOp2.stop();
+      timerNonMaxSuppression3d.stop();
     } else if (_belowLayer_ptr != 0) {
       // Check below.
-      TimerSwitchable timerFancyOp2(
+      TimerSwitchable timerNonMaxSuppression3d(
           "0.3 BRISK Detection: 3d nonmax suppression (per layer)");
       std::vector<typename ScoreCalculator_t::PointWithScore> pt_tmp;
       pt_tmp.reserve(points.size());
@@ -320,7 +356,7 @@ void ScaleSpaceLayer<SCORE_CALCULATOR_T>::detectScaleSpaceMaxima(
         pt_tmp.push_back(*it);
       }
       points.assign(pt_tmp.begin(), pt_tmp.end());
-      timerFancyOp2.stop();
+      timerNonMaxSuppression3d.stop();
     }
 
   }
@@ -1522,3 +1558,5 @@ __inline__ float ScaleSpaceLayer<SCORE_CALCULTAOR_T>::subpixel2D(
       + coeff3 * delta_x + coeff4 * delta_y + coeff5 * delta_x * delta_y
       + coeff6) / 18.0;
 }
+}  // namespace brisk
+#endif  // BRISK_INTERNAL_SCALE_SPACE_LAYER_INL_H_
