+++ brisk_ros_demo/src/livedemo.cc
@@ -1,34 +1,60 @@
 /*
- * main.cpp
- *
- *  Created on: 15.12.2010
- *      Author: lestefan
+ Copyright (C) 2011  The Autonomous Systems Lab, ETH Zurich,
+ Stefan Leutenegger, Simon Lynen and Margarita Chli.
+
+ Copyright (C) 2013  The Autonomous Systems Lab, ETH Zurich,
+ Stefan Leutenegger and Simon Lynen.
+
+ BRISK - Binary Robust Invariant Scalable Keypoints
+ Reference implementation of
+ [1] Stefan Leutenegger,Margarita Chli and Roland Siegwart, BRISK:
+ Binary Robust Invariant Scalable Keypoints, in Proceedings of
+ the IEEE International Conference on Computer Vision (ICCV2011).
+
+ This file is part of BRISK.
+
+ All rights reserved.
+
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions are met:
+     * Redistributions of source code must retain the above copyright
+       notice, this list of conditions and the following disclaimer.
+     * Redistributions in binary form must reproduce the above copyright
+       notice, this list of conditions and the following disclaimer in the
+       documentation and/or other materials provided with the distribution.
+     * Neither the name of the <organization> nor the
+       names of its contributors may be used to endorse or promote products
+       derived from this software without specific prior written permission.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+ DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <sys/time.h>
-#include <ros/ros.h>
-#include <image_transport/image_transport.h>
-#include <opencv2/core/core.hpp>
-#include <opencv2/features2d/features2d.hpp>
-#include <opencv2/highgui/highgui.hpp>
-#include <opencv2/opencv.hpp>
-#include <boost/bind.hpp>
 #include <boost/thread/thread.hpp>
 #include <boost/thread/mutex.hpp>
-#include <boost/shared_ptr.hpp>
-//#include "../include/descriptorEval/leuteneggerDescriptor.h"
+
 #include <brisk/brisk.h>
+#include <brisk/brute-force-matcher.h>
+#include <brisk/internal/brisk-layer.h>
+
+#include <image_transport/image_transport.h>
+#include <opencv2/features2d/features2d.hpp>
 #include <opencv2/nonfree/features2d.hpp>
-#include <opencv2/nonfree/nonfree.hpp>
-#include <fstream>
-#include <iostream>
-#include <brisk/HarrisScoreCalculator.hpp>
-#include <brisk/ScaleSpaceFeatureDetector.hpp>
+#include <opencv2/highgui/highgui.hpp>
+#include <ros/ros.h>
 #include <sm/timing/Timer.hpp>
 
 using namespace std;
 
-const int NUM_REF_IMG = 1;  //
+const int NUM_REF_IMG = 1;
 
 ///DEBUG
 cv::Mat outimg;
@@ -303,7 +329,7 @@ class BriskVisualizer {
             // adapt display
             tmp.push_back(cv::Mat(image->rows / 2, image->cols / 2, CV_8U));
             if (sharedData.numRefImages == 2)
-              cv::BriskLayer::halfsample(*image, tmp.back());
+              brisk::BriskLayer::halfsample(*image, tmp.back());
             else
               tmp.push_back(*image);
             //cv::imshow("BRISK Demo",tmp);
